Module ReWeb

ReWeb - an ergonomic web framework. Start by looking at Server for an overview of the framework. See bin/Main.re for an example server.

See https://github.com/yawaramin/re-web/ for sources.

module Cache: { ... };

Concurrent safe caching.

module Client: { ... };

Make web requests.

module Config: { ... };

Configure and override the behaviour of ReWeb and applications built with it.

module Form: { ... };

Encode and decode web forms to/from specified types.

module Header: { ... };

Create correct response headers.

module Response: { ... };

Send responses.

module Server: { ... };

Create and serve endpoints.

module Topic: { ... };

Subscribe to messages from publishers, and publish messages for subscribers.

module Filter = Server.Filter;

Transform requests and responses in the pipeline.

module Request = Server.Request;

Read requests.

module Service = Server.Service;

Services model the request-response pipeline.