Module type Cache.S

Aside from make and access, the rest of the functions in this interface are similar to the ones in Hashtbl, so you can use those as a reference.

type key;

module Table: Stdlib.Hashtbl.SeededS with type Table.key = key;

type t('a);

let access: t('a) => (Table.t('a) => 'b) => Lwt.t('b);

access(cache, f) runs the function f on the hash table contents of the cache, returning the result value. It can be used to both read from and write to the cache.

f(table) is a callback which has locked access to the hash table contained in the cache and can do anything with it, including multiple operations like checking if the cache contains a key, then adding it or not.

Since this does lock the cache, it is preferable to not do long-running operations here, and to finish and unlock as quickly as possible.

let add: t('a) => key:key => 'a => Lwt.t(unit);

let find: t('a) => key:key => Lwt.t('a);

let find_opt: t('a) => key:key => Lwt.t(option('a));

let iter: t('a) => f:(key => 'a => unit) => Lwt.t(unit);

let length: t('a) => Lwt.t(int);

let make: unit => t('a);

make() allocates a new cache value.

let mem: t('a) => key:key => Lwt.t(bool);

let remove: t('a) => key:key => Lwt.t(unit);

let reset: t('a) => Lwt.t(unit);