Module Client.New

Make requests with a one-shot i.e. stateless client.

type request_body('resp) = ?⁠config:config => ?⁠headers:headers => ?⁠body:Piaf.Body.t => string => Lwt_result.t([> Response.http ] as 'resp, string);

The type of request functions which send a request body.

type request_nobody('resp) = ?⁠config:config => ?⁠headers:headers => string => Lwt_result.t([> Response.http ] as 'resp, string);

The type of request functions which don't send a request body.

let delete: request_body([> Response.http ]);

let get: request_nobody([> Response.http ]);

let head: request_nobody([> Response.http ]);

let patch: request_body([> Response.http ]);

let post: request_body([> Response.http ]);

let put: request_body([> Response.http ]);

let request: ?⁠config:config => ?⁠headers:headers => ?⁠body:Piaf.Body.t => meth:Piaf.Method.t => string => Lwt_result.t([> Response.http ], string);

Make a request not covered by the above functions.