Module ReWeb.Config

Configure and override the behaviour of ReWeb and applications built with it.

let string: string => option(string);

string(name) gets a value from the system environment variable with the name REWEB__name. The following functions all use the same name prefix and work for their specific types.

let bool: string => option(bool);

let char: string => option(char);

char(name) gets the first character of the value of the system environment variable named REWEB__name.

let float: string => option(float);

let int: string => option(int);

module type S = { ... };

The known ReWeb configuration settings.

module Default: S;

Default values for ReWeb configuration settings.