Module ReWeb.Header

Create correct response headers.

module AccessControlAllowOrigin : sig ... end

module CacheControl : sig ... end

See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control for detailed explanations of the various cache options. See ReWeb.Filter.cache_control for the intended usage of this type.

module ContentSecurityPolicy : sig ... end

Convenience module for creating a Content-Security-Policy header. Note that to set up a reporting endpoint properly, you will need to use ReWeb.Filter.csp.

module Directive : sig ... end

module SetCookie : sig ... end

See https://developer.mozilla.org/en-US/docs/Web/HTTP/Cookies for explanations of all the options below.

module StrictTransportSecurity : sig ... end

Convenience to create the HSTS header.