Module Server.Service

module Config : sig ... end

module Request : Request.S with type Reqd.t = Httpaf.Reqd.t with type 'ctx t = 'ctx Request.t

type ('ctx, 'resp) t = 'ctx Request.t -> 'resp Lwt.t

A service is an asynchronous function from a request to a response.

type 'ctx all = ('ctx, [ Response.http | Response.websocket ]) t

A type modelling services with an intersection of all response types. This type is most useful for services which set response headers, or come after filters which set response headers.