Module type ReWeb.Filter

Transform services. Please see here for documentation.

The filters here that don't read the request body additionally preserve whatever context was already in the request before the current filter ran. They do this by putting the previous context in a prev method in the new context object. This is just a convention but a useful one.

The filters which read the body don't do this because they only work with request which have no context (i.e., context of type unit).

module Service: Service.S;

type t('ctx1, 'ctx2, 'resp) = Service.t('ctx2, 'resp) => Service.t('ctx1, 'resp);

A filter transforms a service. It can change the request (usually by changing the request context) or the response (by actually running the service and then modifying its response).

Filters can be composed using function composition.

let basic_auth: t('ctx1, {. username: string, password: string, prev: 'ctx1, }, Response.t(_));

basic_auth decodes and stores the login credentials sent with the Authorization header or returns a 401 Unauthorized error if there is none.

let bearer_auth: t('ctx1, {. bearer_token: string, prev: 'ctx1, }, Response.t(_));

bearer_auth stores the bearer token sent with the Authorization header or returns a 401 Unauthorized error if there is none.

let body_form: Form.t('ctor, 'ty) => t(unit, 'ty, [> Response.http ]);

body_form(typ) is a filter that decodes a web form in the request body and puts it inside the request for the next service. The decoding is done as specified by the form definition typ. If the form fails to decode, it short-circuits and returns a 400 Bad Request.

let body_json: t(unit, Ezjsonm.t, [> Response.http ]);

body_json is a filter that transforms a 'root' service (i.e. one with unit context) into a service with a context containing the request body. If the request body fails to parse as valid JSON, it short-circuits and returns a 400 Bad Request.

let body_json_decode: (Ezjsonm.t => Stdlib.result('ty, exn)) => t(Ezjsonm.t, 'ty, [> Response.http ]);

body_json_decode(decoder) is a filter that transforms a service with a parsed JSON structure in its context, to a service with a decoded value of type 'ty in its context. If the request body fails to decode with decoder, the filter short-circuits and returns a 400 Bad Request.

let body_string: t(unit, string, [> Response.http ]);

body_string is a filter that transforms a 'root' service into a service whose context contains the request body as a single string.

let cache_control: Header.CacheControl.t => t('ctx, 'ctx, [ Response.http | Response.websocket ]);

cache_control(policy) is a filter that applies the caching policy policy to the HTTP response.

let multipart_form: typ:Form.t('ctor, 'ty) => (filename:string => string => string) => t(unit, 'ty, [> Response.http ]);

multipart_form(~typ, path) is a filter that decodes multipart form data. typ must be provided but if you don't actually have any other fields in the form you can use Form.empty to decode into an 'empty' (unit) value.

path(~filename, name) is used to get the filesystem absolute path to save the given filename with corresponding form field name. Note that:

The file will be overwritten if it already exists on disk

filename is the basename, not the full path

The filter will short-circuit with a 401 Unauthorized error response if any of the files can't be opened for writing.

This callback gives you a chance to sanitize incoming filenames before storing the files on disk.

let query_form: Form.t('ctor, 'ty) => t('ctx1, {. query: 'ty, prev: 'ctx1, }, Response.t(_));

query_form(typ) is a filter that decodes the request query (the part after the ? in the endpoint) into a value of type 'ty and stores it in the request context for the next service. The decoding and failure works in the same way as for body_form.