Module Filter.Service

module Request: Request.S;

type t('ctx, 'resp) = Request.t('ctx) => Lwt.t('resp);

A service is an asynchronous function from a request to a response.

type http('ctx, 'resp) = t('ctx, [> Response.http ] as resp);

An HTTP service is a service that specifically returns HTTP responses (as opposed to WebSocket).