Module S.Service

module Config: ReWeb__Config.S;

module Request: ReWeb.Request.S;

type t('ctx, 'resp) = Request.t('ctx) => Lwt.t('resp);

A service is an asynchronous function from a request to a response.

type all('ctx) = t('ctx, [ ReWeb.Response.http | ReWeb.Response.websocket ]);

A type modelling services with an intersection of all response types. This type is most useful for services which set response headers, or come after filters which set response headers.