Module Topic.Make

Make(Msg) is a module for topics that can handle messages of type Msg.t.

Parameters

Msg: { ... }

Signature

type subscription;

type t;

let close: t => Lwt.t(unit);

close(topic) closes an open topic. It has no effect on a closed topic.

let make: unit => t;

make() is a new topic. Typically you will create these at a scope that can pass them to any parts of the application that need them.

let publish: t => msg:Msg.t => Lwt.t(unit);

publish(topic, ~msg) publishes msg onto the topic. This broadcasts the msg to all subscribers of the topic.

let stream: subscription => Lwt_stream.t(Msg.t);

stream(subscription) is a live stream of all messages published to the topic subscribed to with subscription.

let subscribe: t => Lwt.t(subscription);

subscribe(topic) is a subscription to the topic.

let unsubscribe: subscription => Lwt.t(unit);

unsubscribe(subscription) unsubscribes from the topic that was subscribed to with subscription.