Module ReWeb__Cache

The cache module here uses Lwt to ensure that caches are accessed serially to prevent inconsistent data accesses.

module type S = { ... };

Aside from make and access, the rest of the functions in this interface are similar to the ones in Hashtbl, so you can use those as a reference.

module InMemory: (Key: Stdlib.Hashtbl.SeededHashedType) => S with type key = Key.t;

Create a module to manage a concurrent in-memory cache.

module IntKey: Stdlib.Hashtbl.SeededHashedType with type IntKey.t = int;

Use this to create a cache module for caches with int keys.

module StringKey: Stdlib.Hashtbl.SeededHashedType with type StringKey.t = string;

Use this to create a cache module for caches with string keys.