Module ReWeb__Form

Use make (see below) to create a www form that can be decoded to custom types using the form validation rules defined in fields. Use encode to encode a value as a form.

type decoder('a) = string => Stdlib.result('a, string);

A decoder is a function that takes an encoded value and returns a result of decoding the value.

module Field: { ... };

Allows creating a list of form fields using normal list syntax with a local open (e.g. Field.[bool("remember-me"), string("username")]

type t('ctor, 'ty);

A web form is a list of fields and a 'constructor' that takes their decoded field values and returns a value of type 'ty.

let decode: t('ctor, 'ty) => list((string, list(string))) => Stdlib.result('ty, string);

let decoder: t('ctor, 'ty) => decoder('ty);

decoder(form) takes a form definition (see above) and returns a decoder from that form to a type 'ty.

let empty: t(unit, unit);

empty is an empty form i.e. one that does not decode any form data and always succeeds at it.

let encode: ('ty => list((string, string))) => 'ty => string;

encode(fields, value) is a query-encoded string form. It calls fields value to get the representation of value as a list of key-value string pairs.

let make: Field.list('ctor, 'ty) => 'ctor => t('ctor, 'ty);

make(fields, ctor) allows creating a form that can be used to decode (with decoder) www forms.