Module ReWeb__Form.Field

Allows creating a list of form fields using normal list syntax with a local open (e.g. Field.[bool("remember-me"), string("username")]

type t('a);

A form field is a field name and a decoder function for the field.

type list(_, _) =

| ([]) : list('a, 'a)

| (::)(t('a), list('b, 'c)) : list('a => 'b, 'c)

;

Used to create a form field list.

let bool: string => t(bool);

bool(name) returns a boolean field named name.

let ensure: ('a => bool) => t('a) => t('a);

ensure(pred, field) returns a field that will succeed decoding a value a of type 'a if pred(a) is true. Otherwise it will fail decoding.

let float: string => t(float);

float(name) returns a float field named name.

let int: string => t(int);

int(name) returns an integer field named name.

let make: string => decoder('a) => t('a);

make(name, decoder) returns a field with a decoder of type 'a.

let string: string => t(string);

string(name) returns a string field named name.