Module Service.Request

module Config: Config.S;

module Reqd: ReWeb.Request.REQD with type t('fd, 'io) = Httpaf.Reqd.t('fd, 'io);

type t('ctx) = Request.t('ctx) = {

ctx: 'ctx,
Any value. Can be changed by filters. Most useful if it's an object type so that filters can arbitrarily put named values of any type in the request-response pipeline.

query: string,
Either the query string (anything following the ? in the path) or empty string.

reqd: Reqd.t(Lwt_unix.file_descr, Lwt.t(unit)),

};

let body: t(unit) => ReWeb.Body.t;

body(request) gets the request body. There is a chance that the body may already have been read, in which case trying to read it again will error. However in a normal request pipeline as bodies are read by filters, that should be minimized.

let body_string: ?⁠buf_size:int => t(unit) => Lwt.t(string);

body_string(?buf_size, request) returns the request body converted into a string, internally using a buffer of size buf_size with a default configured by ReWeb.Config.S.buf_size.

let context: t('ctx) => 'ctx;

let cookies: t(_) => list((string, string));

let header: string => t(_) => option(string);

header(name, request) gets the last value corresponding to the given header, if present.

let headers: string => t(_) => list(string);

headers(name, request) gets all the values corresponding to the given header.

let make: string => Reqd.t(Lwt_unix.file_descr, Lwt.t(unit)) => t(unit);

make(query, reqd) returns a new request containing the given query and Httpaf reqd.

let meth: t(_) => Httpaf.Method.t;

meth(request) gets the request method (`GET, `POST, etc.).

let query: t(_) => string;

query(request) gets the query string of the request.