Module S.Request

module Config: ReWeb__Config.S;

module Reqd: ReWeb.Request.REQD;

type t('ctx);

A request. 'ctx is the type of the request context which can be updated by setting a new context. It is recommended to do so in a way that preserves the old context (if there is one), e.g. inside an OCaml object with a prev method that points to the old context.

let body: t(unit) => Piaf.Body.t;

body(request) gets the request body. There is a chance that the body may already have been read, in which case trying to read it again will error. However in a normal request pipeline as bodies are read by filters, that should be minimized.

let body_form_raw: ?⁠buf_size:int => t(unit) => Lwt_result.t(list((string, list(string))), string);

body_form_raw(?buf_size, request) returns the request body form decoded into an association list, internally using a buffer of size buf_size with a default configured by ReWeb.Config.S.buf_size.

since

0.7.0

let body_string: ?⁠buf_size:int => t(unit) => Lwt.t(string);

body_string(?buf_size, request) returns the request body converted into a string, internally using a buffer of size buf_size with a default configured by ReWeb.Config.S.buf_size.

let context: t('ctx) => 'ctx;

let cookies: t(_) => list((string, string));

let header: string => t(_) => option(string);

header(name, request) gets the last value corresponding to the given header, if present.

let headers: string => t(_) => list(string);

headers(name, request) gets all the values corresponding to the given header.

let make: string => Reqd.t => t(unit);

make(query, reqd) returns a new request containing the given query and Httpaf reqd.

let meth: t(_) => Httpaf.Method.t;

meth(request) gets the request method (`GET, `POST, etc.).

let query: t(_) => string;

query(request) gets the query string of the request. This is anything following the ? in the request path, otherwise an empty string.

let set_context: 'ctx2 => t('ctx1) => t('ctx2);

set_context(ctx, request) is an updated request with the given context ctx.