Module ReWeb

ReWeb - an ergonomic web framework. Start by looking at Server for an overview of the framework. See bin/Main.re for an example server.

See https://github.com/yawaramin/re-web/ for sources.

module Cache : sig ... end

Concurrent safe caching.

module Client : sig ... end

Make web requests.

module Config : sig ... end

Configure and override the behaviour of ReWeb and applications built with it.

module Form : sig ... end

Encode and decode web forms to/from specified types.

module Header : sig ... end

Create correct response headers.

module Response : sig ... end

Send responses.

module Server : sig ... end

Create and serve endpoints.

module Topic : sig ... end

Subscribe to messages from publishers, and publish messages for subscribers.

module Filter = Server.Filter

Transform requests and responses in the pipeline.

module Request = Server.Request

Read requests.

module Service = Server.Service

Services model the request-response pipeline.