Module Client.New

Make requests with a one-shot i.e. stateless client.

type 'resp request_body = ?config:config -> ?headers:headers -> ?body:Piaf.Body.t -> string -> ([> Response.http ] as 'resp, string) Lwt_result.t

The type of request functions which send a request body.

type 'resp request_nobody = ?config:config -> ?headers:headers -> string -> ([> Response.http ] as 'resp, string) Lwt_result.t

The type of request functions which don't send a request body.

val delete : [> Response.http ] request_body

val get : [> Response.http ] request_nobody

val head : [> Response.http ] request_nobody

val patch : [> Response.http ] request_body

val post : [> Response.http ] request_body

val put : [> Response.http ] request_body

val request : ?config:config -> ?headers:headers -> ?body:Piaf.Body.t -> meth:Piaf.Method.t -> string -> ([> Response.http ], string) Lwt_result.t

Make a request not covered by the above functions.